[#13] refactor: 테이블뷰 datasource 분리
Closes #13
